Fix disconnecting  Remote as follower when no leader is attached (#17856)
Fixes #17822

Summary of the issue:
An error occurs when disconnecting Remote when in follower mode if no leader is attached.

Description of user facing changes
No error occurs.

Description of development approach
The error occurs because, when a follower session disconnects, it makes sure that the secure desktop handler no longer holds a reference to it. The setter for that reference, however, attempts to mutate it. Since we're setting it to None, this fails.

Thus, in secureDesktop.SecureDesktopHandler's followerSession setter, check that the new session is not None before accessing its properties.