Only play sounds or beeps for Remote cues (#17852)
Fixes #17849

Summary of the issue:
If set to play beeps instead of sounds, Remote still plays the sound, as well as the beep, when outputting cues.

Description of user facing changes
Only beeps or sounds are played, as appropriate.

Description of development approach
Fixed the logic in remoteClient.cues._playCue to have the sound and beep playing code on separate branches.

Previously, if playing sounds was selected, the sound was played and the function returned before the beep code was reached. However, this also stopped text cues from being output.
When the premature return was removed, the logic was not updated, resulting in both a beep and sound being heard when beep output was selected.