Libre Office: Fix German heading formatting commands (#17836)
Fix-up of #17153.

Summary of the issue:
When modifying heading style with ctrl+1/2/3/4/5 or turning back to body text with ctrl+0, other formatting changes might be reported in languages where "ctrl" is translated differently, e.g. "Strg" in German.

Description of user facing changes
Only style and no other associated formatting changes should be reported when modifying heading styles with ctrl+1/2/3/4/5 or turning back to body text with ctrl+0.

Description of development approach
Do not use the displayName property to define which type of gesture we are dealing with, Use modifierNames and mainKeyName properties instead since they are not translated.