Verify screen curtain at runtime (#17894)
Fixes #12708
Supercedes #12701

Summary of the issue:
There is a security risk if the screen curtain fails silently. The Magnification API used by Screen Curtain does not officially support WOW64 applications like NVDA. As this is a risk with untested Windows versions, an external check would be helpful to minimise silent failures of the screen curtain.

Notwithstanding this, as this is a security feature, it is better to check that it is working as expected, and inform the user if it is not.

Description of user facing changes
None.

Description of development approach
Implemented a new function , isScreenFullyBlack, in NVDAHelper/local/screenCurtain.cpp. This method:

Obtains a reference to the desktop window, which covers the entire virtual screen, and thus all monitors connected to the computer.
USES GDI to capture the screen.
USES GDIPlus to calculate a histogram of this screen capture.
While this seems like overkill, it performs significantly better on my machine than individually checking that each pixel is black, likely due to optimisations and hardware accelleration in GDIPlus.
Checks that, for each channel of the histogram, the value at 0 is the area of the screen.
Call this function after applying the fullscreen colour effect and hiding the cursor in visionEnhancementProviders.screenCurtain.ScreenCurtainProvider.__init__. If the return is False, raise RuntimeError, which disables the screen curtain and informs the user that screen curtain failed to activate.