Rationalise Remote gestures (#17865)
Fixes #17803

Summary of the issue:
The current gesture for switching control between the local and remote
machines conflicts with the SysTrayList add-on, which is popular among
those coming from JFW.

Description of user facing changes
Changed this gesture to be `NVDA+alt+tab`, which I feel is more
intuitive.

Also unbound the "Push clipboard" command, as no suitable, easily
pressed combination was immediately obvious, and this command is
available from the NVDA menu.

Description of development approach
Changed the script decorators.

Testing strategy:
Ran from source and checked that the new bindings work.

Known issues with pull request:
None.