Check before saving Add-on Store settings (#17920)
Fixes #17805
Follow-up to #17597

Summary of the issue:
An error is logged when exiting from the launcher, as
`addonStore.storeSettings._AddonStoreSettings.save` is called regardless
of the return value of `NVDAState.shouldWriteToDisk`.

Description of user facing changes
An error tone is no longer heard after running the launcher.

Description of development approach
Only attempt to save Add-on Store settings if
`NVDAState.shouldWriteToDisk` returns `True`.

While the error was harmless, it was worrying some users.

Testing strategy:
Created a portable copy with the launcher for NVDA alpha 35962,a503ae32
to ensure the issue was present.

Created a launcher (`scons launcher`), and used it to create a portable
copy. Verified that the issue no-longer occurs.

Known issues with pull request:
None.