Merge pull request #17883 from nvaccess/remotePushClipboard
Fixes #17881
Fixes #17875

Summary of the issue:
Currently, the push clipboard command:

Fails with a logged error but no useful output when attempting to push a non-text clipboard
Reports success when there is no-one else connected.
Description of user facing changes
An error message is reported when attempting to push non-text clipboard
An error message is reported when pushing the clipboard to an empty channel
Description of development approach
Catch OSError when attempting to push clipboard, as this is raised when there is no text content on the clipboard.
Record connected leaders and followers on both leader and follower sessions. Add read-only properties, connectedLeadersCount, connectedFollowersCount, and connectedClientsCount (sum of leaders and followers counts) to RemoteSession.
Add a read only property, connectedClientsCount to RemoteClient, which grabs the connectedClientsCount from the appropriate session, or returns 0 if there is no session in progress.
Check that the remote client's count of connected clients is greater than 0 before pushing the clipboard.