Remote, log warning instead of error on secure desktop when no IPC file exists (#17924)
Summary of the issue:
When no follower session connected to a system, the system doesn't
create an IPC file when opening the secure desktop. As the secure
desktop copy of NVDA tries to find the IPC and logs an exception when it
can't, this results on an error on the secure desktop.
Now this error is safely ignored when in secure mode, it is still
noticeable when setting serviceDebug to 1.

Description of user facing changes
None

Description of development approach
Change exception (error) into a warning.

Testing strategy:
Can't test on secure desktop, but code review should be sufficient here.

Known issues with pull request:
None known