Don't send send keys gesture if not connected (#17927)
Fixes #17874

Summary of the issue:
Performing the "Toggle Control" gesture (NVDA+alt+tab) when no Remote Access session is in progress results in the gesture being sent to the operating system.

Description of user facing changes
The gesture is always trapped by NVDA.

Description of development approach
In remoteClient.client.RemoteClient.toggleRemoteKeyControl, no longer call gesture.send if there is no leader session. Also inform the user if not connected, if not connected as leader, or if there are no followers connected.