Fix syntax for clickable URL of GPL v2 license (#17936)
Summary of the issue:
During 2025.1 dev cycle, the licence has become an HTML document, converted from markdown, and URL are clickable. In markdown < and > (angle brackets) make an URL clickable.
Though, the URL of GPL v2 is between < and />, as if it was an HTML tag; this leads to an extra slash in the URL when the license is displayed, causing an Error 404 when clicked.

Description of user facing changes
When clicked in the license, the GPL v2 page now correctly opens in the browser.