Fix NVDA not reading the ribbon in Geekbench 6.4 (#17937)
Fixes #17892

Summary of the issue:
NVDA Does not read the controls in the Geekbench 6.4 ribbon when navigating and errors in the log.

Description of user facing changes
The menu can be navigated and interacted with normally.

Description of development approach
Geekbench uses something similar to the Microsoft Office ribbon. The code to check for older Office versions expected a normal productVersion for Office apps, however, Geekbench's productVersion is "6,4,0,0". This is the cause of the ValueError. I've fixed this by only checking the version if isOfficeApp is True first.

Testing strategy:
Tested that the Geekbench ribbon can be navigated and interacted with.
Tested the Microsoft Office ribbon just in case.