Fix redundant table end markers in tables in Word with UIA enabled (#17938)
Fixes #15828

Summary of the issue:
When UIA is enabled in word and the cursor is halfway a table cell, NVDA will braille table end ath the cursor position.

Description of user facing changes
Table end will no longer be shown.

Description of development approach
In cases where this happens, it looks like there are at leass two findings.

textRange.CompareEndpoints returns a huge number (always 6029412 in my case)
It does so regardless where comparing start to start, end to end
This also means that the number of the end to end comparison is positive, basically meaning that the end of the inner range is greater than the end of the outer range. This should be impossible anyway.