Move crowdin sync and mozilla symbols from the legacy deploy script (which never runs) to pushPackagingInfo at end of build. (#17955)
Summary of the issue:
Crowdin is not getting the latest source strings for nvda.pot.
When we moved to the new deploy webhook on the new NV Access Infrastructure, there was an assumption that the old deploy script would still run. This is not the case. If a deploy section is specified, deploy_script does not run. There were some extra actions in the old deploy script such as uploading to Crowdin and the Mozilla symbol server, but as the old deploy script no longer runs, these actions do not occur.

Description of user facing changes
None.

Description of development approach
Moved upload to Crowdin and Mozilla Symbol Server from the old deployScript.ps1 and into pushPackagingInfo.ps1, which always runs in on_finish.