Merge pull request #17943 from nvaccess/remoteTerminology
Fixes #17815

Summary of the issue:
NVDA's new Remote Access feature currently uses inconsistent terminology.

Description of user facing changes
Remote Access:

Is consistently referred to as "Remote Access".
Uses the term "computer" to refer to clients.
Previously both "computer" and "machine" were used. I also considered "device", as this seems to be the preferred terminology in Windows.
Uses the word "send" instead of "push" for sending the clipboard to connected clients.
Has descriptive help items for its menu items, all of which are complete sentences.
Uses sentence fragments for UI messages where appropriate
Description of development approach
Generated a pot file, and searched for the word "remote" therein. If it was referring to Remote Access, and some other name was used, replaced it with Remote Access.

Searched for the word "remote" in the user guide and edited as appropriate. Also did a full readthrough of the remote section.

Reviewed the strings in the settings dialog, menu, and cues explicitly and edited as appropriate.

Calculated the diff 1b4a28d4d...9833d7812. Searched in it for newly introduced translator comments, and edited them if related to Remote Access and I thought they were unclear.

Calculated the diff 9833d7812...42aaea791. Reviewed it for any changed strings that used gettext (_(...)), and changed to use pgettext (pgettext("remote", ...)).