Remove unnecessary reference to Python test module (#18013)
Fixes #18005 

Summary of the issue:
Remote client has an unnecessary reference to the Python test module. This module is meant as a module internal to Python testing. The reference is unnecessary because the mocket socket is never used within the test.

Description of user facing changes
None

Description of development approach
Just removed the patch call