Allow Remote Access leaders to regain control after the last follower has disconnected (#17970)
Fixes #17965

Summary of the issue:
If the controlling computer is sending keyboard input to the controlled
computer when the controlled computer disconnects from the session, the
controlling computer is unable to regain control of the local session.

Description of user facing changes
* The `NVDA+Alt+Tab` gesture now allows control to return to the local
computer even if no followers are connected.

Description of development approach
* In the `toggleRemoteKeyControl` method of `RemoteClient`, adjust the
threshhold for refusing to toggle sending keys from simply having
connected followers, to currently controlling the local computer and
having connected followers.
* This means that the check does not apply when sending keys, meaning
control can be returned to the local computer regardless of the number
of connected followers.
* Add new `_switchToRemoteControl` and `_switchToRemoteControl` methods
to `RemoteClient`, and refactor `toggleRemoteKeyControl` to use these
new methods. The logic in these methods is almost identical to that in
the original `toggleRemoteKeyControl` method.

Testing strategy:
In a Remote Access channel with a leader and a follower:

* Switch the leader to controlling the follower. Perform some actions on
the leader to ensure it is working as expected. Return the leader to
controlling the current computer.
* Switch the leader to controlling the follower. Disconnect the follower
from the session. Switch the leader back to local control.
* Attempt to switch the leader to controlling the follower. Observe that
it fails.

Known issues with pull request:
If the user is controlling a remote computer when that computer
disconnects, no indication is given that subsequent keyboard input is
(essentially) a no-op. In attempting to add such an indication, I
discovered #18004.