Always focus remote connection dialog (#18050)
Fixes #18047

Summary of the issue:
If the Remote Access connection dialog is the first NVDA GUI to be shown, it is not automatically focused.

Description of user facing changes
The Remote Access connection dialog is automatically focused when opened, even if it is the first NVDA GUI to be shown.

Description of development approach
Add a wx.EVT_SHOW handler that manually brings the dialog to the foreground and focuses it.