Use MathML attributes for PDFs read in Adobe Acrobat (#17984)
This fixes #17980

Summary of the issue:
For PDF, the code was not grabbing MathML attributes needed for speech.

Description of user facing changes
The speech for math, and rarely the braille (e.g., bevelled fractions in Nemeth), was not always correct due to attributes not being picked up and the defaults being used.

Description of development approach
Unfortunately, the PDF interface does not allow grabbing all the attributes. Instead, one must ask for each attribute individually. Most attributes don't affect speech or braille, so it is not necessary to get them. I looked at what MathCAT used and added those.