Save previously connected Remote Access servers to disc (#18057)
Fixes #18037

Summary of the issue:
Remote Access's previously connected servers aren't being written to disc when the config is saved.

Description of user facing changes
Previously connected Remote Access servers are now persisted across NVDA restarts.

Description of development approach
The issue is because configobj only recognises config items as having changed when __setitem__ is called. Since lists are mutable, this was not happening.

Tidied up the _remoteClient.configuration.writeConnectionToConfig function a bit.
Don't take any action if the server to persist is already the last server stored in config.conf["remote"]["connections"]["lastConnected"].
Set config.conf["remote"]["connections"]["lastConnected"] to itself so that configobj recognises that it has changed, and will thuss write it out when saving the config.