Improve logging in Remote Access (#18061)
Summary of the issue:
Remote Access's logging strategy is currently suboptimal.

Description of user facing changes
A new debug logging category, remoteClient, has been added.
When this is enabled, more information is logged by Remote Access.
Log messages that were previously not stored are now shown in NVDA's logs.
Description of development approach
In the config schema, add a new boolean key, remoteClient, to the debugLogging category. Add this to the list of log categories in the advanced settings panel. Also add a _isDebugForRemoteClient function to _remoteClient.configuration which returns the value of this config item.
In the TCPTransport class, if debug logging is enabled for remoteClient, log incoming and outgoing messages to make debugging easier.
In _remoteClient.localMachine and _remoteClient.serializer, switch to using logHandler.log, rather than a custom logger, so that all logs are collected in the same place.