Make Remote Access settings visible but read only in secure mode (#18071)
Closes #17946
Supersedes #17987

Summary of the issue:
Remote Access's settings are entirely absent from the settings dialog in
secure mode.

Description of user facing changes
Remote Access's settings are now present, but disabled, in the settings
dialog in secure mode.

Description of development approach
Add all controls in the Remote Access settings panel to a static box
sizer so that they can be disabled en masse. If running in secure mode,
disable this static box. Also add a `wx.adv.BannerWindow` explaining why
the controls are read-only, similar to the Add-on Store dialog. Move
`RemoteSettingsPanel` to always be present in the settings categories.

Testing strategy:
Started NVDA normally and in secure mode (with `--secure`) and opened
Remote Access's settings.

Known issues with pull request:
The checkboxes are not visually dimmed, though they are inoperable.