Add externalPythonDependencies logging category (#18067)
Follow up for #10393

Summary of the issue:
Since #10393, we log at level warning or higher for external Python
dependencies. There is no way to enable logging at lower levels, such as
debug. This can however be pretty helpful for external dependencies like
comtypes to see what's going on under the hood.

Description of user facing changes
Have extra logging when a new debugging category,
`externalPythonDependencies`, is enabled.

Description of development approach
The overall logging level is now determined from the level of the root
logger, which is no longer limited to a minimum of WARNING. Instead, a
filter is implemented on the handler that filters appropriately. Note
that initially, it made sense to add this filter to the root logger, but
filters don't propagate from the root logger to all loggers below.

Testing strategy:
Calling `comtypes.CoInitialize` with and without the new category
enabled, logging at level debug. Ensured that behavior was as expected
(i.e. comtypes logged when the category was enabled)

Known issues with pull request:
In #10393, I mentioned issues with re-entrancy when comtypes logs at
level debug. I have not yet reproduced this. That said, the logging
categories are documented in such a way that they should only be enabled
on request. I don't think this needs extra documentation.