Only show Remote Access actions when available (#18068)
Fixes #18039

Summary of the issue:
The mute and send SAS items are available in the Remote Access submenu
even when connected as a follower.

Description of user facing changes
These options are only shown when connected as a leader.

Description of development approach
In `_remoteClient.menu.Menu.handleConnected`, only enable the mute and
send SAS menu items if connected and connection mode is leader.
In `_remoteClient.client.RemoteClient`, updated the `toggleMute` and
`sendSAS` methods to report failure and return early when not connected
as leader.
Updated unit tests to work with the new checks.

Testing strategy:
Manually tested that the items are unavailable in the given situations.
Unit tests.

Known issues with pull request:
None.