Support quick nav to headings 7 to 9 in browse mode  (#18015)
Fixes #18014

Summary of the issue:
In MS Word, heading levels can go from 1 to 9 but only quick nav keys for 1 to 6 are available (as on the web).

Description of user facing changes
Quick navigation keys 1 to 9 are now available in any browsable document. In MS Word, headings 1 to 9 are supported natively. In HTML, only H1 to H6 are supported, but level 7 to 9 can be implemented throught aria-level.

Description of development approach
Created a function to add heading scripts to a class, taking a level range as parameter
Used this function to factorize heading 1 to 9 quick nav keys for browseable documents.